Firebase and AppsFlyer integrations, more features
Integrated firebase crashlytics and performance monitoring. Integrated Appsflyer SDK successfully to start advertising. Added color-coded messages for each person so it is easier to keep track of who is who. Added video quality indicator in the Levels page. Added toggle to disable usage of cellular data when downloading videos (Will default to built in media). Now an option to not require audio in video playback.