fixed microphone problem
After contacting Apple technical and a week of back and forth, we were able to pinpoint of the microphone turning on after returning to the foreground. Turns out, the background music needed to be paused each time the game went into the background. A simple "willEnterBackground" in the AppDelegate was able to fix this issue. After months of scouring, finally the issue is no more!