Downloading Efficiency upgrades, UI upgrades
Downloading content is way faster for the user and there are better checks in place to make sure the user has the right content. Error messages in place in case videos are not available or are corrupted etc.

Few more UI updates such as adding motion shadows to the toolbars in each level.