Fixed iPad bugs, other features
iPad does not support action sheets without a popover, so I had to add a couple lines to make it compatible with iPad.

Created a message blast option to send message notification to everyone (no need for OneSignal). I might remove OneSignal later, although I should keep it as a backup.

Made notifications easier to manage and was picky about when the alerts should show, as they glitched things up. When notification tapped, it redirects to ChatViewController (MessageView for ADMIN).