API removal, Localization finalization, Trend fixes
OneSignal and AppsFlyer are no longer integrated with the app!! This means no more "Ask app not to track".

Localization code is now updated and works well. The game downloads data from firebase, creates a file, which is then used as the primary source file for localization. I am also reducing the size of the backup localization files before the download actually occurs.

As for trend fixes, some strings were lost in translation or badly formatted, so I am trying to make it as decent as possible.